Calculate openingMessage directly in OpeningHoursProvider instead of on demand
